Enable LZMA compression for arm64 kernel Image with Depthcharge
The arm64 kernel Image files are not compressed as part of the kernel
build.  It can be done separately and it is up to the bootloader to
handle decompression.  As Depthcharge supports LZMA for arm64 kernels,
enable this feature to significantly reduce the size of the resulting
FIT image.

This is especially useful to keep the Depthcharge TFTP buffer size
small while being able to boot kernels that would be too big if
uncompressed.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>