debian/control: make lava-dispatcher Achitecture: all
It was originally changes from all to linux-any, and then tailored by
hand with time. There is no point in keeping it that way, since
1) lava-dispatcher contains no architecture-specific code, 2) it now
depends on systemd-sysv which is only available on linux-any anyway so
being Linux-only is still the case.

With this, it's also required to drop the architecture-specific
qualifiers from Recommends.