udev: fix race condition by iterating on the devices
Without that patch, the action will wait forever if the udev event was
triggered before the action has time to listen on the socket.

The timing issue is fixed by:
* starting the monitor
* listing the devices currently available
* listen for events

Every udev events that will be generated after the start of the monitor will be
queued are handled when listening for events.

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>