fastboot: always wait for device to show up
When fastboot runs on the dispatcher host or on a lxc container where
the device is going to be shared with it, fastboot itself will wait for
the device to appear. But to share the device with a one-off docker
container run, we need it to be available before we start the container.

For consistency, let's wait for the device to appear before each
fastboot deploy.