doc: actions-test: Improve intro to the types of test actions.
The main ideas of the changes:
1. Make it explicit that there're 3 different types of test actions
available.
2. Primary way to identify a test action type should be whatever is
used in job definition YAML (that's what user will see both at the
start and the end). So, make explict the mapping between YAML directives
and "descriptive names".

The rest of changes to clarify, make more explicit, and more generic
of the wording, to make it more suitable for novices with very different
usecases in mind.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>