doc: actions-test: interactive: Clarify/explicitisize description
This is largely formatting/punctuation changes, but also tries to
make more explicit various usage patterns/implications. For example,
"if LAVA matches one of the prompt and ``successes`` is defined, an
error will be recorded. If ``successes`` is not defined, then matching
a prompt will generate a passing result." is absolutely correct
"mathematical" description, which is largely useless for a novice,
because neither reasons for that behavior, nor implications of it are
immediately clear. So, instead make it explicit that this behavior allows
to reliably handle interactive commands which produce some output on
success, and also handle commands which don't produce any output on
success.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>