Drop u-boot commands if the placeholder is not replaced
When booting with u-boot and nfs, providing a ramdisk is optional. Anyway LAVA
was always sending the command to load the ramdisk. In recent u-boot versions,
a command to load a non-exisring tftp resource will fail.

With this patch, the command will not be sent if the placaholder is not
replaced (because the coresponding value is None).

Closes #347

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>