doc: interactive: Write down a complete reference for current behavior
This is based both on a new adopter's questions ("how exactly it works?" -
"oh, some aspects are not described"), then actual experience using it,
looking at the source code to clarify some aspects, and then implementing
missing features/clarifying behavior. It aims to provide fairly complete
description of each directive, and enough description of the overall
behavior. Stylistic-wise, following changes were made: a) There was visible
overuse of note's and warning's. They are useful to interleave a normal
narrative from time to time, but if there 2 notes and 2 warnings in row,
that's hardly productive and just interferes with reader's attention.
Most notes/warnings are merged into description of the corresponding
options. b) As some options are inter-dependent, don't hesitate to
repeat some aspects (as this is reference, we should expect a user
to just look up whatever they think they need now, instead of re-read
the entire section). c) This is a reference again, but don't make it
too dry, and leave "absolutely correct and absolutely useless" impression.
Describe some implications and give some hints. Don't get carried away
though, specific case studies should be described in usage guide, not
a reference.

And there's a usage guide ("Writing an interactive test action"), so
crosslink the reference and it. (The usage itself needs updating, not
done in this commit.)

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>