qemu: delay the substitution
Generate and use the substitutions dictionary during the run() phase.

Using and setting namespace data in the validate pahse is never a good idea as
this might be overriden by another action later on.
Better to use namspace data only during the run phase.

This will help to fix an issue when using archived resources.

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>