Add OrderedDict YAML representer
This is to prevent Lava to include `object/apply` constructs such as:
  !!python/object/apply:collections.OrderedDict
in the yaml files made available through XMLRPC endpoints.

This became more problematic recently when PyYAML version 5.2 was
released which made the usage of the `load()` method raise an error when
using `object/apply` as it can be used as an attack vector. Here is the
error I am getting:
  yaml.constructor.ConstructorError: could not determine a constructor
  for the tag 'tag:yaml.org,2002:python/object/apply:collections.OrderedDict'
    in "<unicode string>", line 1, column 11:
      {"extra": !!python/object/apply:collection ...

Because Lava produces yaml files with `object/apply` constructs PyYAML
5.2 fails to parse them using the regular `load()` method. We have to
use the compatibility `unsafe_load()` method.

I believe producing such constructs is not on purpose because we can see
a commit [1] preventing the use of `object/apply` for another Python
type.

This commit adds a representer to convert the OrderedDict to yaml's
Ordered Mapping.

[1] https://github.com/Linaro/lava/commit/14b347c51efb89e8d6c64d9aa74fb4190f03227c

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>