Rework FVP to work with !1004 overlay format
New way of directing where to put the normal LAVA test shell overlay
binaries, which works in a similar way to the new !1004 overlays change.

Instead of specifying some modules tar ball to add as an overlay,
specify that you are placing "lava" overlay and then set how to append
that overlay (ramdisk image, partition of a disk image, guest image etc).

Change-Id: Ida56e175d1cd74537b0219a5e2f42469f7bbb612
Signed-off-by: Dean Birch <dean.birch@arm.com>