dispatcher: deploy: add supporting for postprocessing images with docker
The first piece is a new deployments strategy, "downloads". It will just
download the specified images, and maybe deploy an overlay to them if
that's needed.

    - deploy:
        to: downloads
        # ...
        images:
          rootfs:
            url: http://url.to/image.img

The files are downloaded under the "downloads" subdirectory of the job
artifacts directory, which is not cleaned up after the action is
finished, so that those images can be used later in the job (see below).

The "downloads" strategy also provides a way of postprocessing
downloaded images using user-provided docker container images.  For
that, add a postprocess: entry in the deployment action, specify which
docker image to run, and which commands to run against the images:

    - deploy:
        to: downloads
        # ...
        images:
          rootfs:
            url: http://url.to/image.img
        postprocess:
          docker:
            image: tool
            steps:
              - tool rootfs/stretch.img

The steps are compiled in a shell script that will be executed in a
shell inside a docker container running the specified docker image. The
current working directory is the top-level directory where images are
downloaded. Any modifications made to the images, or even new files
created under the downloads directory, are persisted so that the
following actions can use them.

The last piece is the new "downloads://" URL scheme can not be used to
refer to those files downloaded previosly, or even to files created
there via postprocessing of downloaded images.

For example, suppose you want to merge a rootfs image with a kernel
image. One can first download the original images:

    - deploy:
        to: downloads
        # ...
        images:
          rootfs:
            url: http://url.to/rootfs.img
          kernel:
            url: http://url.to/kernel.img
          modules:
            url: http://url.to/modules.tar.gz
        postprocess:
            docker:
                image: mytool
                steps:
                    - mytool-inject-modules --rootfs rootfs.img --output system.img

One can then use the image created in the previous step with any
deployment strategy:

    - deploy:
        to: WHATEVER
        images:
            kernel:
                url: downloads://kernel.img
            system:
                url: downloads://system.img