test: interactive: allow for non-exiting commands
When trying to test commands that don't exit in non-posix shells,
interactive test needs a new flag "wait_for_prompt". This flag can be
added to any command. Only success or failure strings need to be matched to
proceed to next command. One use case for this flag is testing
'fastboot' in u-boot shell

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>