dispatcher: docker test shell: pass device to docker container
The lava-dispatcher-host mapping will share the device with the
container when it's connected, so for instance if the tests reset the
board for some reason, it will just work.

But if the device is already up when the container starts, it needs to
be passed in the docker command line.