dispatcher: docker test shell: finalize docker container
This avoids leaving the containers running until the end of the job.

If you have two docker test shell actions and both use adb, and the
first is left running, there will be an adb server blocking the device,
and the second action will never be able to use the devices.