dispatcher: docker test shell: ensure stage is set properly
It turns out the job parser only assigns the correct stage parameters to
tests that declare that they need an overlay. In the case where docker
and target tests are mixed in the same job, this will cause the tests
intended for docker to also end up in the target, but this is probably
not a big deal since they won't be executed anyway.