rewrite share/postinst.py
The script wil now create SECRET_KEY and DATABASES if needed and requested. THe
values will be stored in /etc/lava-server/settings.d/ under 00-secret-key.yaml
and 00-database.yaml.

The procedure to create the database is now using a small SQL file. The need to
create a temp root users is not needed anymore.

The script output has also been improved.

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>