device-types: Add uboot_error_messages variable
The 'uboot_error_messages' variable allows device-type templates to set
their own list of u-boot error messages as required if the default set is
not appropriate. For example it may be required that one of the default
error messages is ignored.

Usage example:
{% set uboot_error_messages = [
  'Resetting CPU',
  'Must RESET board to recover',
  'TIMEOUT'
] %}

The 'uboot_extra_error_message' variable can still be used in conjunction
with 'uboot_error_messages'.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>