UrlRepo: add support tests from url and tar
Since LAVA is not designed as a build farm, it will be used to
work with various build farm together.
So, it will be very useful to support tests/images from different
sources.

In the current LAVA, LAVA supports tests from git/bzr and inline.
Support for tar and url is planned(not usable yet).

Now, with this patch, support for tar and url could be implemented.

Besides, corresponding sample job and unit test are added.

Signed-off-by: Liu Wenlong <liuwl.fnst@cn.fujitsu.com>