Auto-generate devices, device types, aliases and tags from device dictionary.
Add is_synced flag to device model and migration.
This management command will create devices, device types, aliases and tags
based on the device dictionary variable 'sync_to_lava'.
If the device is synced but the file is not present in the filesystem,
the device health will be set to Retired.

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>