Make FIT image format optional with Depthcharge for x86
Depthcharge has support for FIT images on arm/arm64 but not x86.  To
deal with this, make the FIT image conversion optional to only use it
when applicable.  If the "mkimage_fit_arch" parameter is not defined
in the device type, the kernel image is left as-is and not converted
to FIT.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>