Add DEPTHCHARGE_RAMDISK parameter when not using FIT
Depthcharge now supports an extra optional argument with the tftpboot
command to download a separate ramdisk file.  This is useful for
architectures that don't have FIT support in Depthcharge such as x86.
Add a DEPTHCHARGE_RAMDISK for this purpose, to not conflict with the
global RAMDISK parameter which would also be set when using FIT format
and result in the ramdisk to be downloaded twice.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>