Rename Depthcharge FIT parameter to DEPTHCHARGE_KERNEL
The FIT parameter used in Depthcharge device templates assumes the
kernel image is of the FIT format.  While this is normally true for
arm/arm64, it's a misleading name for other architectures such as x86
which don't have FIT support in Depthcharge.  As such, rename it to
DEPTHCHARGE_KERNEL to also not conflict with the global KERNEL
parameter.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>