lava_common/version.py: consider only upstream release tags
This fixes a regression in the APT repo with daily builds, when a
debian/*tag causes the version number from git describe to be smaller
than the version number of the latest daily build created before the
debian/* tag.

85369e53769ce1ea0bf97e7eca3c0bc6fea0c171 fixed producing version numbers
at all in the presence of debian/* tags, but here we are ignoring them
completely.