lava-dispatcher: utils: Add drop_line parameters
The current docstrings of strings.substitute() function refers
to a drop of the line if the value is None. In some cases, the
value is only empty and we may not want to remove the entire line.

To handle these king of use-cases, add a drop_line parameters
that allows to remove the line removal in case a value is None/Empty.

The default behavior is to remove the line, as it was previously done.
To desactivate that, the user must set drop_line parameter to False.

Signed-off-by: Mylène Josserand <mylene.josserand@collabora.com>