cmsis_dap: Add action param "post_unmount_delay", default to 3s.
Allow boards to set "post_unmount_delay" param for "cmsis_dap" boot action,
to wait given number of seconds for system stabilization after "umount"
command. This is prompted by the issues seen on RaspberryPi4 workers, where
some binaries failed to be properly transferred without such delay. In my
limited testing, just 1s delay was enough. But to be on safe side, we set
default to 3s, and allow boards to override it.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>