device-types: Proxy any cmsis-dap parameters from a device dict.
Previously, we had a quite cumbersome way to deploy new boot method
parameters to actual devices:

1. Introduce new param in boot method source code.
2. Add an adhoc, param-specific templating to device-type jinja template.
3. Put actual param setting in a device dict for a particular device.

We'd like to cut the middleman, p.2 from the process. This is also the
most cumbersome step, because primary source of the device-type template
is LAVA upstream. So, if you make a change to an upstream template, you
need to submit changes upstream, yet maintain your local version until
the patch is merged and new LAVA version is released.

This patch introduces a new idiom to pass an arbitrary parameter values
from a device dict to an action. "cmsis-dap" boot action is addressed in
this patch (consistently for all device types which use it).

To set particular params in a device dict, then use snippet like:

{% block cmsis_dap_params %}
post_unmount_delay: 1
foo: bar
{% endblock cmsis_dap_params %}

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>