Do not search qemu binary when docker is in use
When executing a sparc64 qemu job with docker, LAVA throws an
infrastructure error because it do not find the qemu binary.
This is normal since the binary will be in the docker image.

So when actions/boot/qemu is used with the optional docker, dont try to
find the qemu binary.

Signed-off-by: Corentin LABBE <clabbe@baylibre.com>