boot/pyocd: Allow to connect to device before flashing
This will allow to capture serial output appearing immediately after
flashing. Specific reason for this change is lpcxpresso55s69 board,
which doesn't support reset-using-uart-break feature. In other words,
how we dealt re: this problem with other boards is that we flashed it
(and it started executing, which is not good), then connected to it,
which caused reset and executing from start, and that's how we were
able to capture the full output (but potentially with extra output
in the beginning, which is again not good).

That way doesn't work with lpcxpresso55s69 - as there's no reset,
we lose some initial output, and capture only tail of it. This problem
however doesn't happen with "jlink" boot method, and the difference
is exactly that it first issues ConnectDevice sub-action, and then
Flash() action. Looking at the other boot actions, that's also how
most of them do either. "cmsis-dap" and "pyocd" for about the only
clear exceptions.

Given that, it would be tempting to just swap the order unconditionally.
However, detailed gitlog review shows that such a change was already
made on 2018-07-31 in commit 4efd8586a5e8, just to be reverted the
same day in 610b8e58d18. Sadly, the revert doesn't include reasons
for it, by I was able to find the related dicussion at
https://review.linaro.org/c/lava/lava/+/27685. Quoting:

> There's a very specific reason this code was written how it was.
> On some pyocd devices the serial connection doesn't exist until
> the board is flashed, or the flashing process resets the serial
> connection so if already connected it will be closed.
> This change will mean they simply no longer work.

Specific boards weren't named, but it's easily to believe that could
happen, given all the variability in board behavior we see.
So instead (and as suggested in the review discussion by the link
above), a "connect_before_flash" boot method param is introduced,
and the default is kept the same as before (i.e.
connect_before_flash = False). For now, it will be set to True for
lpcxpresso55s69, and then other boards can be tested with that setting
as time permits.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>