dispatcher: docker: support sharing network stacks
Every usage of docker that is backed by lava_dispatcher.utils.docker now
supports sharing networking stacks between container. You need to start
a long-running container, say, a QEMU device that is using docker, and
specify a name for it using the `container_name` key in the job
definition. Later in the job definition, when defining another
container, say a docker test shell one, you reference the name used for
the first container in the `network_from` key. For example:

- In the QEMU boot action:

  docker:
    image: foo/bar
    container_name: qemu

- In a docker test shell after that:

  docker:
    image: something/else
    network_from: qemu

Note that this is only supported if the original container is a
long-running container that stays running in the background after the
action where it started finishes, e.g. a QEMU device container.

We have to drop support for container hostname because that conflicts
with the --network option. Also it's pointless to set the hostname
inside the container.