Remove hard dependency on postgresql on the same system
Downgrade postgresql to a Recommends, making LAVA setup more flexible.
On some installations you may want to have the database on a container
or a on a separate host entirely.

To make that possible we also fix the postinst python script to check
for database availability earlier, while checking for options, so it
can set options.db to False and avoid trying to do any database-related
setup.