Allow overriding configuration through the environment
With this we make it easier to set the database and secret key which are
appropriate for each environment based on the environment files read by
the systemd file and for other use cases, such as LAVA deployments based
on containers.

This change also allows for appending new entries to ALLOWED_HOSTS using
the same method.