dispatcher: boot: deptcharge: Add ResetConnection function
Instead of connecting the device, the ResetConnection function
will be used to perform a reset of the console on retries.
Thanks to that, the console will be reset each time a retry is
performed. On the first try, the disconnet will not be done as
the console is not connected.

Signed-off-by: Mylène Josserand <mylene.josserand@collabora.com>