Make worker token available through manage command
When adding a new worker, everything can be done through the command line
management command. However, you need to go to the Django admin to obtain
the token that you'll need to set up the worker.

This change prints the token upon worker creation and also prints it when
showing details about a worker.