Reintroduce wrongly removed /etc/logrotate.d/lava-scheduler-log
This was a tricky one. What happened:

- The version in stretch had /etc/logrotate.d/lava-scheduler-log
- The version in buster dropped that configuration file, but didn't
  remove it correctly (i.e. it was removed from the filesystem, but was
  still in the dpkg database.
- The version in bullseye adds that file again.

In a stretch -> buster -> bullseye upgrade scenario, the user gets
prompted about changes to the file even if they didn't touch the file
before upgrading to buster, because it was removed incorrectly in the
buster version. We could fix this in buster first, but then the upgraded
would only work if the user is using the very latest stable update
before upgrading.

Instead, on the buster -> bullseye upgrade, if the file is known by dpkg
-S (meaning it was installed by the stretch version) and missing from
the filesystem (i.e. it was removed by the upgrade to buster), we
reinstate the file with its original contents from stretch, so that the
bullseye upgrade will detect it hasn't been changed and just override it
with the new version.

I tested both stretch -> buster -> bullseye and buster -> bullseye
upgrades and both cases are handled correctly.

Closes: #988362