lava_common.compat: add support for PyYAML 5.4
A few tests need to explicitly use an unsafe loader to be able to test
YAML roundtrips. As a fix for a security issue, PyYAML 5.4 requires us
to be even more specific to request an unsafe loader, what is a
backwards-incompatible change.

On older versions of PyYAML that don't provide UnsafeLoader, we need
to fallback to the regular Loader class.

Signed-off-by: Marc Deslauriers <marc.deslauriers@ubuntu.com>
Signed-off-by: Antonio Terceiro <antonio.terceiro@linaro.org>