test: docker: Allow to add arbitrary docker options thru device type
The syntax is based on existing config options for the "deploy" and
"boot" actions in the device type YAML. Example:

actions:
  deploy:
    ...
  boot:
    methods:
      pyocd:
        ...
  test:
    methods:
      docker:
        global_options:
        - --debug
        options:
        - --privileged
        - --mount=type=bind,source=/dev/disk/by-id,destination=/dev/disk/by-id
        - --mount=type=bind,source=/dev/serial/by-id,destination=/dev/serial/by-id
        - {{ docker_shell_extra_arguments }}

As can be seen, there's also provision to set some options via device dict
(`docker_shell_extra_arguments` above). It can be: a) not set (`null` YAML
value, filtered out and not contributing to docker arguments); b) set to a
single string (option value); c) set to a list of strings.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>