frdm-k64f: Add docker-shell config options
With recent changes, we allow to pass additional command-line options
when invoking docker-shell (test: docker: stanza in job definition),
as defined in a particular device type.

Add sample config section showing how to define these options, to both
serve as a reference, and be validated as a part of the testsuite.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>