base-uboot: add uboot_altbank and uboot_altbank_command for NXP LS platforms
On NXP Layerscape boards, we have a mechanism of having multiple banks.
This is done mainly to protect board from getting bricked if wrong
images gets flashed on board.

So here concept of Alternate Bank boot comes to help.
By default board boots from Bank 1, but we will not flash the images
on Bank 1, instead we will flash images on Bank 2.
After flashing the images, we will boot from Bank 2 with a command
on u-boot "qixis_reset altbank" or "cpld reset altbank" depending
on platform.
With this command, Board boots from the Bank 2 and now we can boot
up kernel on board.
Since in this case, we have flashed the images on Bank 2 and even
if images are corrupt, Bank 2 will get bricked, but since we have
set board to boot from Bank 1 by default, Board will still be usable
and you can flash the correct images on Bank 2 and get that working.

This patch is implementing this option by introducing 2 variables:

uboot_altbank: This will indicate that if we want to boot from
alternat bank or not. This variable can be given in context:
section of job. By default it is set as false.

uboot_altbank_cmd: Command which needs to be executed for booting
from alternate bank. On some platforms it is "qixis_reset altbank"
and on some it is "cpld reset altbank"

Also added "uboot_altbank" in CONTEXT_VARIABLES

Signed-off-by: Sahil Malhotra <sahil.malhotra@nxp.com>
Signed-off-by: Rohit Agarwal <rohit.agarwal@nxp.com>