dispatcher: allow root_partition for fastboot deployment
Thanks to u-boot raw partition it's possible to flash a whole
disk image including partition table with fastboot.

In this case we need to define the root partition to deploy
the tests overlay.

I'm just applying this option to non sparse image, because As far
I know it's only possible to have a filesystem image in android sparse
format. (Not disk image)

Signed-off-by: Julien Massot <julien.massot@iot.bzh>