ReadFeedback: For finalize, perform long reads, make duration configurable
Original read-feedback action implementation performed only 1 call to
the underlying pexpect connection (via listen_feedback() method). So,
if pexpect buffered a lot of output, only some initial small chunk of
it (~2K) was returned. And if, say, timeout (duration) for read-feedback
was 20s, and some data was read before that, only that data was returned,
and read-feedback didn't try to wait for the remaining duration (in other
words, the only case when it waited 20s was only when there was no data
in buffer or form input, in other cases it waited less).

This behavior (which we can call "sort reads") is a good match when
reading feedback connections concurrently with the main connection
processing (indeed, we just want to poll feedbacks to *try* to keep
up with the output there, but don't delay main processing too much).
But for read-feedback action, it leads to the fact that a lot of
feedback output will never be processed at all. So, we actually
need to perform "long" reads there, to spool all output which may
be buffered and/or wait for the full duration specified.

Additional complication is however that read-feedback action may be
called not just at the end of job processing, but also at other times.
E.g. the "fvp" boot action calls it after opening feedback connections,
apparently, to bootstrap and synchronize their outputs. The difference
between final vs other read-feedback calls is in "finalize" parameter.

That's exactly what we modify - when read-feedback is executed with
finalize=True parameter, it performs multiple pexpect reads for the
duration specified (unless EOF is seen before that). The behavior
of finalize=False case stays the same, to avoid unexpected change of
behavior in parent actions which use it.

Finally, the question of read-feedback duration. Previously, it was
hardcoded as 1 second. Testing showed that even with the above changes,
not enough feedback output may be captured. That's easy to explain -
as during the main connection processing, we perform only short reads
of feedbacks, of feedback connections produce a lot of info, it
accumulates in buffers, and then the only chance to read it is at the
end of processing (read-feedbacks(finalize=True)), and that may take
its time too (including more than 1s).

This duration also depends on particular test payload, and not on the
device, in other words, this duration must be specified on the level
of job, not on the level of device dictionary. It's however complication
by the fact that read-feedback action isn't specified in jobs explicitly,
but rather added in sub-pipelines to higher-level actions. Thus, parameters
(like duration) can't be specified explicitly. With some stretch,
read-feedback duration can be called read-feedback timeout. And that's
exactly where this patch adds, it can be specified in job's top-level
"timeouts" section:

timeouts:
  connections:
    read-feedback: 20

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>