boot/fvp: Consume all feedback from an FVP binary
The issue is similar to one previously fixed for the "read-feedback"
action: only a single "short" read was performed. This in particular
led to loss of code coverage reports printed out by the FVP binary
(if requested with command-line options).

Instead, read all data while it comes (i.e. until EOF or timeout is
hit). Unlike previous fix for the "read-feedback" action, there's
no detailed timeout handling: a hardcoded while of 5s is used as
before, and for the case that and FVP binary will produce too much
output (not an expected situation), we rely on general job timeout
to stop it.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>