test: interactive: Log prompt match after successes/failures
We logged prompt match for an interactive test script without
successes/failures, but didn't log it in case a success/failure
was matched (even though we also match a prompt afterwards in
this case too).

Be consistent and log matched prompt in all cases. This makes
LAVA logs self-describing and easier to understand.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>