docker + adb,fastboot: avoid race condition when device disconnects
We often observe race conditions where a USB device will disconnect
between the time when it gets added to the docker command line as a
--device= option and the time that the container actually starts.
Instead of running one-off containers, start the containers first, wait
for them to be up, map devices into them with lava-dispatcher-host, and
then run the command. This way, if the device gets disconnected it will
be re-shared with the container once it is enumerated again.