DeviceContainerMappingMixin: allow failures when triggering udev
This avoids a race condition when a device gets disconnected between the
moment it is listed by get_udev_devices and the time that `udevadm
trigger` is called. If the `udevadm trigger` call fails,
lava-dispatcher-host might still share the device with the container
when triggered by udev itself when the device reconnects.