lava_dispatcher: fix case insensitive urlparse effect
When using urlparse the resulting object has 'hostname' method. This
method always returns lowercase strings. In case of deploy-to-downloads
this might lead to issues with finding files that have names with
uppercase characters. For example this section will result in an error:
- deploy:
    to: flasher
    images:
        image:
            url: downloads://FOO-BAR.img

To fix this 'hostname' should be replaced with 'netloc' which is case
sensitive.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@foundries.io>